// src/types/translationTypes.d.ts
type classes = {
    title: string;
    filters: {
        language: string[];
        levels: string[];
        dayTime: string[];
        reset: string;
    };
    body: {
        title: string;
        notMatch: {
            title: string;
            info: string;
        };

    };
    card: {
        book: string;
        details: string;
        join: string;
    };
};
// ...
type description = {
    class: {
        items: Array<string>,
        buttons: Array<string>
    }, teacher: {
        titles: Array<string>,
        buttons: Array<string>
    }

}

type profileNav = {
    nav: {
        find: string,
        balance: string,
        suscribe: string,
        refer: string,
    },
    optionsNav: {
        home: string,
        messages: string,
        myClasses: string,
        booked: string,
        purchase: string,
        myLessons: string,
        settings: string
        logOut: string
    }
}

type profile = {
    home: {
        title: string,
        hoursTaken: string,
        classesTaken: string,
        storiesTime: string,

    },
    messages: {
        all: string,
        unread: string,
        archived: string
    },
    student: {

        lesson: string
    },
    teacher: {


        home: string,
        classes: {
            title: string
            add: string,
            table:
            {
                created: string,
                name: string,
                language: string,
                level: string,
                date: string,
                duration: string
            },
            options: {
                edit: string,
                cancel: string
            }
        }
        purchase: string,
        form: {
            name: string,
            duration: string,
            language: string,
            link: string,
            level: string,
            price: string,
            date: string,
            capacity: string,
            description: string,
            button: string,
            back: string
        }

    }

}